feat: Adicionado verificação de owner ao realizar ações

+++ back/src/main/java/br/com/airbnb/service/UsuarioService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -13,6 +14,7 @@ import br.com.airbnb.domain.usuario.Foto;
 import br.com.airbnb.domain.usuario.Usuario;
 import br.com.airbnb.repository.FotoUsuarioRepository;
 import br.com.airbnb.repository.UsuarioRepository;
+import br.com.airbnb.service.exception.NaoPossuiPermissaoAlterarException;
 import br.com.airbnb.service.image.ImageResponse;
 import br.com.airbnb.service.image.ImageService;
 
@@ -36,12 +38,19 @@ public class UsuarioService {
 
 	@Transactional
 	public Usuario uploadImagem(Long id, MultipartFile file) {
+
 		Optional<Usuario> optional = this.usuarioRepository.findById(id);
 		if (!optional.isPresent()) {
 			throw new EntityNotFoundException();
 		}
 
 		Usuario usuario = optional.get();
+
+		Usuario usuarioContext = (Usuario) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		if (!usuarioContext.getId().equals(usuario.getId())) {
+			throw new NaoPossuiPermissaoAlterarException();
+		}
+
 		if (usuario.getFoto() != null) {
 			Foto foto = usuario.getFoto();
 			this.fotoUsuarioRepository.delete(foto);
